refactor: 에러 핸들링 로직을 Http11Processor에서 Servlet으로 이동
